# openapi.yaml
openapi: 3.0.3
info:
  title: Dynamic Car Sharing API
  description: |-
    A comprehensive API for a peer-to-peer car sharing platform.
    This API covers all functionalities for renters, car owners, and administrators,
    from Browse and booking cars to managing disputes and verifying users.
  version: 1.0.0
servers:
  - url: /api/v1
    description: API Version 1
...
paths:
...
  /bookings:
    get:
      tags:
        - Bookings
      summary: List all bookings (Admin)
      description: Retrieves a paginated list of all bookings on the platform. For admin use.
      operationId: getAllBookings
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: A paginated list of all bookings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingPage'
    post:
      tags:
        - Bookings
      summary: Book a car
      description: Allows a renter to book a car for a specific time period.
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreateRequest'
      responses:
        '201':
          description: "Booking request created."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/BadRequest'
  /bookings/quote:
    post:
      tags:
        - Bookings
      summary: Get booking quote
      description: Calculates estimated booking price before creating a booking.
      operationId: getBookingQuote
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Optional key to return the same quote response for repeated retries.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequest'
      responses:
        '200':
          description: Quote calculated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /bookings/waitlist:
    post:
      tags:
        - Bookings
      summary: Join booking waitlist
      description: Adds renter to waitlist for an already occupied car time range.
      operationId: joinBookingWaitlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingWaitlistCreateRequest'
      responses:
        '201':
          description: Waitlist entry created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingWaitlistEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
  /bookings/waitlist/{waitlistId}:
    get:
      tags:
        - Bookings
      summary: Get active waitlist entry
      operationId: getBookingWaitlistEntry
      parameters:
        - name: waitlistId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Active waitlist entry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingWaitlistEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
  /bookings/waitlist/{waitlistId}/cancel:
    patch:
      tags:
        - Bookings
      summary: Cancel waitlist entry
      operationId: cancelBookingWaitlistEntry
      parameters:
        - name: waitlistId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Waitlist entry canceled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingWaitlistEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
  /bookings/availability:
    get:
      tags:
        - Bookings
      summary: Check car availability for a time range
      description: Returns whether the car can be booked for the given interval without creating a reservation. Uses the same rules as booking creation for car status and overlapping PENDING or APPROVED bookings.
      operationId: checkCarAvailability
      parameters:
        - name: carId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: startTime
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Availability result; HTTP 400 if the time range is invalid or the car cannot be resolved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarAvailabilityResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /bookings/availability/calendar:
    get:
      tags:
        - Bookings
      summary: Get daily car availability calendar
      description: Returns per-day availability for a single car in the provided time range. Each day is marked as available or unavailable with a reason.
      operationId: getCarAvailabilityCalendar
      parameters:
        - name: carId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: startTime
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Daily availability calendar for the requested car and interval.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarAvailabilityCalendarResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /bookings/{bookingId}:
    get:
      tags:
        - Bookings
      summary: Get booking details
      description: Retrieves details of a specific booking.
      operationId: getBookingById
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Detailed information for the booking.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
        - Bookings
      summary: Update booking status
      description: Allows a car owner or renter to update the booking status (e.g., APPROVE, CANCEL).
      operationId: updateBookingStatus
      parameters:
        - $ref: '#/components/parameters/BookingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/TransactionStatus'
      responses:
        '200':
          description: "Booking status updated."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      tags:
        - Bookings
      summary: Delete a booking (Admin)
      description: Allows an admin to delete a booking record.
      operationId: deleteBooking
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '204':
          description: "Booking deleted successfully."
        '404':
          $ref: '#/components/responses/NotFound'
  /bookings/{bookingId}/summary:
    get:
      tags:
        - Bookings
      summary: Get booking summary with payment and transactions
      description: Returns booking details together with the linked payment (if any) and all transactions for that booking in one response.
      operationId: getBookingSummary
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Aggregated booking, optional payment, and transactions for the booking.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingSummary'
        '204':
          description: No booking exists for the given identifier.
...
components:
  parameters:
...
    BookingId:
      name: bookingId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    Page:
      name: page
      in: query
      description: "Page number (0-indexed)"
      schema:
        type: integer
        default: 0
    Size:
      name: size
      in: query
      description: "Number of items per page"
      schema:
        type: integer
        default: 20
    Sort:
      name: sort
      in: query
      description: "Sorting criteria: property,(asc|desc)"
      schema:
        type: string
        example: "createdAt,desc"
  responses:
    NotFound:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    BadRequest:
      description: Invalid request payload.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    Booking:
      $ref: 'schemas/Booking.yaml'
...
    Payment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bookingId:
          type: integer
          format: int64
        amount:
          type: number
          format: double
        cancellationPenaltyAmount:
          type: number
          format: double
        cancellationRefundAmount:
          type: number
          format: double
        status:
          $ref: '#/components/schemas/TransactionStatus'
        paymentMethod:
          $ref: '#/components/schemas/PaymentType'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Transaction:
      $ref: 'schemas/Transaction.yaml'
    BookingSummary:
      type: object
      required:
        - booking
        - transactions
      properties:
        booking:
          $ref: '#/components/schemas/Booking'
        payment:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Payment'
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
...
    BookingCreateRequest:
      $ref: 'schemas/BookingCreateRequest.yaml'
...
    BookingWaitlistCreateRequest:
      type: object
      required:
        - renterId
        - carId
        - startTime
        - endTime
        - pickupLocationId
      properties:
        renterId:
          type: integer
          format: int64
        carId:
          type: integer
          format: int64
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        pickupLocationId:
          type: integer
          format: int64
        promoCode:
          type: string
    BookingWaitlistEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
        renterId:
          type: integer
          format: int64
        carId:
          type: integer
          format: int64
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        pickupLocationId:
          type: integer
          format: int64
        promoCode:
          type: string
        status:
          type: string
          enum:
            - ACTIVE
            - CANCELED
            - PROMOTED
        createdAt:
          type: string
          format: date-time
    QuoteRequest:
      type: object
      required:
        - renterId
        - carId
        - startTime
        - endTime
      properties:
        renterId:
          type: integer
          format: int64
        carId:
          type: integer
          format: int64
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        pickupLocationId:
          type: integer
          format: int64
        promoCode:
          type: string
        loyaltyPointsToUse:
          type: number
          format: double
    QuoteResponse:
      type: object
      properties:
        renterId:
          type: integer
          format: int64
        carId:
          type: integer
          format: int64
        baseAmount:
          type: number
          format: double
        dynamicMarkupAmount:
          type: number
          format: double
        discountAmount:
          type: number
          format: double
        loyaltyAmount:
          type: number
          format: double
        totalAmount:
          type: number
          format: double
        currency:
          type: string
        expiresAt:
          type: string
          format: date-time
    CarAvailabilityResponse:
      type: object
      required:
        - available
      properties:
        available:
          type: boolean
        reason:
          type: string
          enum:
            - CAR_NOT_AVAILABLE
            - SLOT_TAKEN
          description: Present when available is false.
    CarAvailabilityCalendarDay:
      type: object
      required:
        - date
        - available
      properties:
        date:
          type: string
          format: date
        available:
          type: boolean
        reason:
          type: string
          enum:
            - CAR_NOT_AVAILABLE
            - SLOT_TAKEN
    CarAvailabilityCalendarResponse:
      type: object
      required:
        - carId
        - days
      properties:
        carId:
          type: integer
          format: int64
        days:
          type: array
          items:
            $ref: '#/components/schemas/CarAvailabilityCalendarDay'
...
    BookingPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Booking'
        totalPages:
          type: integer
        totalElements:
          type: integer
...
    TransactionStatus:
      type: string
      enum: [ PENDING, APPROVED, COMPLETED, CANCELED, REFUNDED ]
...
    PaymentType:
      type: string
      enum: [ CREDIT_CARD, DEBIT_CARD, PAYPAL, APPLE_PAY, GOOGLE_PAY, BANK_TRANSFER, CASH, CRYPTO, MOBILE_MONEY ]
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string